requirements.txt
GeoAlchemy2~=0.14.7
marshmallow~=3.21.1
flask~=3.0.3
geojson~=3.1.0
apispec~=6.6.0
python-dotenv~=1.0.1
SQLAlchemy~=2.0.29
flask_sqlalchemy~=3.1.1
flask_marshmallow~=0.15.0
marshmallow_sqlalchemy~=0.29.0
shapely
psycopg2-binary
pydantic_settings
gunicorn~=21.2.0
flask_apispec